requirements.txt
asgiref==3.8.1
Django==5.0.3
packaging==24.0
pillow==10.2.0
pytesseract==0.3.10
sqlparse==0.4.4
tzdata==2024.1
